refactor, move each graphic class into module
