+++ fastplotlib/graphics/__init__.py
@@ -1,2 +1,6 @@
-from .graphics import Image, Scatter, Line, Histogram
+from .histogram import Histogram
+from .line import Line
+from .scatter import Scatter
+from .image import Image
+
 __all__ = ["Image", "Scatter", "Line", "Histogram"]
