+++ fastplotlib/graphics/histogram.py
@@ -1,0 +1,75 @@
+from _warnings import warn
+from typing import Union, Dict
+
+import numpy as np
+import pygfx
+
+from fastplotlib.graphics._base import Graphic
+
+
+class _HistogramBin(pygfx.Mesh):
+    def __int__(self, *args, **kwargs):
+        super(_HistogramBin, self).__init__(*args, **kwargs)
+        self.bin_center: float = None
+        self.frequency: Union[int, float] = None
+
+
+class Histogram(Graphic):
+    def __init__(
+            self,
+            data: np.ndarray = None,
+            bins: Union[int, str] = 'auto',
+            pre_computed: Dict[str, np.ndarray] = None,
+            colors: np.ndarray = None,
+            draw_scale_factor: float = 100.0,
+            draw_bin_width_scale: float = 1.0
+    ):
+
+        if pre_computed is None:
+            self.hist, self.bin_edges = np.histogram(data, bins)
+        else:
+            if not set(pre_computed.keys()) == {'hist', 'bin_edges'}:
+                raise ValueError("argument to `pre_computed` must be a `dict` with keys 'hist' and 'bin_edges'")
+            if not all(type(v) is np.ndarray for v in pre_computed.values()):
+                raise ValueError("argument to `pre_computed` must be a `dict` where the values are numpy.ndarray")
+            self.hist, self.bin_edges = pre_computed["hist"], pre_computed["bin_edges"]
+
+        self.bin_interval = (self.bin_edges[1] - self.bin_edges[0]) / 2
+        self.bin_centers = (self.bin_edges + self.bin_interval)[:-1]
+
+        # scale between 0 - draw_scale_factor
+        scaled_bin_edges = ((self.bin_edges - self.bin_edges.min()) / (np.ptp(self.bin_edges))) * draw_scale_factor
+
+        bin_interval_scaled = scaled_bin_edges[1] / 2
+        # get the centers of the bins from the edges
+        x_positions_bins = (scaled_bin_edges + bin_interval_scaled)[:-1].astype(np.float32)
+
+        n_bins = x_positions_bins.shape[0]
+        bin_width = (draw_scale_factor / n_bins) * draw_bin_width_scale
+
+        self.hist = self.hist.astype(np.float32)
+
+        for bad_val in [np.nan, np.inf, -np.inf]:
+            if bad_val in self.hist:
+                warn(f"Problematic value <{bad_val}> found in histogram, replacing with zero")
+                self.hist[self.hist == bad_val] = 0
+
+        data = np.vstack([x_positions_bins, self.hist])
+
+        super(Histogram, self).__init__(data=data, colors=colors, colors_length=n_bins)
+
+        self.world_object: pygfx.Group = pygfx.Group()
+
+        for x_val, y_val, bin_center in zip(x_positions_bins, self.hist, self.bin_centers):
+            geometry = pygfx.plane_geometry(
+                width=bin_width,
+                height=y_val,
+            )
+
+            material = pygfx.MeshBasicMaterial()
+            hist_bin_graphic = _HistogramBin(geometry, material)
+            hist_bin_graphic.position.set(x_val, (y_val) / 2, 0)
+            hist_bin_graphic.bin_center = bin_center
+            hist_bin_graphic.frequency = y_val
+
+            self.world_object.add(hist_bin_graphic)
